Test add task
