+++ src/modules/task.js
@@ -1,3 +1,5 @@
+// import render from './display.js';
+
 export default class Tasks {
   constructor(description) {
     this.description = description;
@@ -16,59 +18,22 @@ export default class Tasks {
         task.index = i + 1;
       });
       localStorage.setItem('array', JSON.stringify(array));
-      document.querySelector('.message').innerHTML = '*Added new task successfully';
-      this.render();
-      window.location.reload();
-      document.querySelector('#add-task').value = '';
     }
   }
 
   removeTask(index) {
+    this.index = index;
     const array = JSON.parse(localStorage.getItem('array')) || [];
     const newArray = array.filter((e) => e.index !== index);
     newArray.forEach((task, i) => {
       task.index = i + 1;
     });
     localStorage.setItem('array', JSON.stringify(newArray));
-    this.render();
   }
 
   editTask(index, newDescription) {
     const taskToEdit = this.find((task) => task.index === index);
     taskToEdit.description = newDescription;
     localStorage.setItem('array', JSON.stringify(this));
-  }
-
-  render() {
-    const array = JSON.parse(localStorage.getItem('array')) || [];
-    const toDo = document.querySelector('.to-do-items');
-    toDo.innerHTML = '';
-    if (array.length > 0) {
-      for (let i = 0; i < array.length; i += 1) {
-        toDo.innerHTML += `
-        <div class="list-items" data-index="${array[i].index}">
-          <div class="task-container">
-            <input class="checkbox-tick" type="checkbox">
-            <input class="checkbox" class="desc" value="${array[i].description}">
-          </div>
-          <button class="btn-more remove"><i class="fa-solid fa-ellipsis-vertical"></i></button>
-        </div>`;
-      }
-    }
-
-    const checkboxes = document.querySelectorAll('.checkbox-tick');
-    for (let i = 0; i < array.length; i += 1) {
-      checkboxes[i].checked = array[i].completed;
-    }
-
-    // add event listener to remove button
-    const removeButtons = document.querySelectorAll('.remove');
-    removeButtons.forEach((button) => {
-      button.addEventListener('click', (event) => {
-        event.preventDefault();
-        const index = parseInt(event.target.parentElement.parentElement.getAttribute('data-index'), 10);
-        this.removeTask(index);
-      });
-    });
   }
 }
